restrict editing when auction is over

+++ app/api/auctionlistings.py
@@ -24,12 +24,12 @@ def get_auction_listing(auctionlisting_id):
         return {"errors": "Auctionlisting not found."}, 404
 
     if request.method == "DELETE":
-        # if not single_listing.check_owner(owner_id):
-        #     return {"errors": "Forbidden."}, 403
-        # else:
-        db.session.delete(single_listing)
-        db.session.commit()
-        return {"Success": "Listing deleted."}, 202
+        if not single_listing.check_owner(single_listing.owner_id):
+            return {"errors": "Forbidden."}, 403
+        else:
+            db.session.delete(single_listing)
+            db.session.commit()
+            return {"Success": "Listing deleted."}, 202
 
     if request.method == "PUT":
         form = AuctionListingForm()
